break out RemoteSyncSettingsForm component into smaller chunks
