+++ enterprise/frontend/src/metabase-enterprise/remote_sync/components/LibrarySyncRow/LibrarySyncRow.tsx
@@ -3,20 +3,17 @@ import { t } from "ttag";
 
 import CS from "metabase/css/core/bordered.module.css";
 import { Box, Flex, Icon, Switch, Text } from "metabase/ui";
-import type { RemoteSyncConfigurationSettings } from "metabase-types/api";
 
 import { SYNC_LIBRARY_PENDING_KEY } from "../../constants";
-
-type FormValues = RemoteSyncConfigurationSettings & {
-  [SYNC_LIBRARY_PENDING_KEY]?: boolean;
-};
+import type { RemoteSyncSettingsFormState } from "../../types";
 
 interface LibrarySyncRowProps {
   isReadOnly: boolean;
 }
 
 export const LibrarySyncRow = ({ isReadOnly }: LibrarySyncRowProps) => {
-  const { values, setFieldValue } = useFormikContext<FormValues>();
+  const { values, setFieldValue } =
+    useFormikContext<RemoteSyncSettingsFormState>();
   const isChecked = values[SYNC_LIBRARY_PENDING_KEY] ?? false;
 
   const handleToggle = (checked: boolean) => {
