+++ enterprise/frontend/src/metabase-enterprise/remote_sync/components/RemoteSyncAdminSettings/CollectionsToSyncSection.tsx
@@ -1,0 +1,32 @@
+import { t } from "ttag";
+
+import { Stack, Text } from "metabase/ui";
+
+import { SharedTenantCollectionsList } from "../SharedTenantCollectionsList";
+import { TopLevelCollectionsList } from "../TopLevelCollectionsList";
+
+import { RemoteSyncSettingsSection } from "./RemoteSyncSettingsSection";
+import { useSetting } from "metabase/settings";
+
+export const CollectionsToSyncSection = () => {
+  const useTenants = useSetting("use-tenants");
+
+  return (
+    <RemoteSyncSettingsSection
+      title={t`Collections to sync`}
+      description={t`Choose which collections to sync with git.`}
+    >
+      <Stack gap="lg">
+        <TopLevelCollectionsList />
+        {useTenants && (
+          <>
+            <Text fw={700} size="md" lh="1rem">
+              {t`Shared collections`}
+            </Text>
+            <SharedTenantCollectionsList />
+          </>
+        )}
+      </Stack>
+    </RemoteSyncSettingsSection>
+  );
+};
