+++ enterprise/frontend/src/metabase-enterprise/remote_sync/types.ts
@@ -1,5 +1,7 @@
 import type { State } from "metabase/redux/store";
+import type { RemoteSyncConfigurationSettings } from "metabase-types/api";
 
+import type { SYNC_LIBRARY_PENDING_KEY } from "./constants";
 import type { SyncTaskState } from "./sync-task-slice";
 
 export interface RemoteSyncStoreState extends State {
@@ -7,3 +9,9 @@ export interface RemoteSyncStoreState extends State {
     remoteSyncPlugin?: SyncTaskState;
   };
 }
+
+export type RemoteSyncSettingsVariant = "admin" | "settings-modal";
+
+export type RemoteSyncSettingsFormState = RemoteSyncConfigurationSettings & {
+  [Key in typeof SYNC_LIBRARY_PENDING_KEY]?: boolean;
+};
